Refactor child management and attendance modules; remove unused classes and update relationships to use UUIDs
