+++ src/main/java/com/careconnect/coreapi/billing/jpa/Billing.java
@@ -1,12 +1,9 @@
 package com.careconnect.coreapi.billing.jpa;
 
-import com.careconnect.coreapi.childmgmt.jpa.Guardian;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
 
 import java.math.BigDecimal;
 import java.time.Instant;
@@ -23,10 +20,8 @@ public class Billing {
     @Column(name = "id", nullable = false, updatable = false)
     private UUID id;
 
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @OnDelete(action = OnDeleteAction.CASCADE)
-    @JoinColumn(name = "guardian_id", nullable = false)
-    private Guardian guardian;
+    @Column(name = "guardian_id", nullable = false)
+    private UUID guardianId;
 
     @Column(name = "amount", precision = 65, scale = 30)
     private BigDecimal amount;
