+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/AllergyService.java
@@ -1,11 +1,11 @@
-package com.careconnect.coreapi.childmgmt.service;
+package com.careconnect.coreapi.childmgmt.internal;
 
-import com.careconnect.coreapi.childmgmt.jpa.Allergy;
-import com.careconnect.coreapi.childmgmt.jpa.Child;
-import com.careconnect.coreapi.childmgmt.jpa.ChildAllergy;
-import com.careconnect.coreapi.childmgmt.repository.AllergyRepository;
-import com.careconnect.coreapi.childmgmt.repository.ChildAllergyRepository;
-import com.careconnect.coreapi.childmgmt.repository.ChildRepository;
+import com.careconnect.coreapi.childmgmt.domain.Allergy;
+import com.careconnect.coreapi.childmgmt.domain.Child;
+import com.careconnect.coreapi.childmgmt.domain.ChildAllergy;
+import com.careconnect.coreapi.childmgmt.internal.AllergyRepository;
+import com.careconnect.coreapi.childmgmt.internal.ChildAllergyRepository;
+import com.careconnect.coreapi.childmgmt.internal.ChildRepository;
 import com.careconnect.coreapi.common.exceptions.DuplicateResourceException;
 import com.careconnect.coreapi.common.exceptions.ResourceNotFoundException;
 import com.careconnect.coreapi.common.exceptions.ValidationException;
