+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/ChildAllergyRepository.java
@@ -1,6 +1,6 @@
 package com.careconnect.coreapi.childmgmt.internal;
 
-import com.careconnect.coreapi.childmgmt.jpa.ChildAllergy;
+import com.careconnect.coreapi.childmgmt.domain.ChildAllergy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
