+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/ChildGuardianService.java
@@ -1,11 +1,11 @@
-package com.careconnect.coreapi.childmgmt.service;
+package com.careconnect.coreapi.childmgmt.internal;
 
-import com.careconnect.coreapi.childmgmt.jpa.Child;
-import com.careconnect.coreapi.childmgmt.jpa.ChildGuardian;
-import com.careconnect.coreapi.childmgmt.jpa.Guardian;
-import com.careconnect.coreapi.childmgmt.repository.ChildGuardianRepository;
-import com.careconnect.coreapi.childmgmt.repository.ChildRepository;
-import com.careconnect.coreapi.childmgmt.repository.GuardianRepository;
+import com.careconnect.coreapi.childmgmt.domain.Child;
+import com.careconnect.coreapi.childmgmt.domain.ChildGuardian;
+import com.careconnect.coreapi.childmgmt.domain.Guardian;
+import com.careconnect.coreapi.childmgmt.internal.ChildGuardianRepository;
+import com.careconnect.coreapi.childmgmt.internal.ChildRepository;
+import com.careconnect.coreapi.childmgmt.internal.GuardianRepository;
 import com.careconnect.coreapi.common.exceptions.DuplicateResourceException;
 import com.careconnect.coreapi.common.exceptions.ResourceNotFoundException;
 import com.careconnect.coreapi.common.exceptions.ValidationException;
