--- src/main/java/com/careconnect/coreapi/childmgmt/jpa/Enrollment.java
@@ -1,52 +1,0 @@
-package com.careconnect.coreapi.childmgmt.jpa;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-import java.time.Instant;
-import java.util.UUID;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "enrollments")
-public class Enrollment {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @ColumnDefault("gen_random_uuid()")
-    @Column(name = "id", nullable = false, updatable = false)
-    private UUID id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @OnDelete(action = OnDeleteAction.CASCADE)
-    @JoinColumn(name = "child_id", nullable = false)
-    private Child child;
-
-    @Column(name = "start_date")
-    private Instant startDate;
-
-    @Column(name = "end_date")
-    private Instant endDate;
-
-    @Column(name = "status", length = Integer.MAX_VALUE)
-    private String status;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "enrolled_by_guardian_id")
-    private ChildGuardian enrolledBy;
-
-    @Column(name = "admin_sign_off_id")
-    private UUID adminSignOffId;
-
-    @ColumnDefault("CURRENT_TIMESTAMP")
-    @Column(name = "created_at", nullable = false)
-    private Instant createdAt;
-
-    @Column(name = "updated_at", nullable = false)
-    private Instant updatedAt;
-
-}
