+++ src/main/java/com/careconnect/coreapi/communications/domain/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Domain entities for the communications module.
+ * 
+ * This package contains the core domain entities that represent messaging concepts.
+ * These entities should not be directly accessed by other modules.
+ */
+package com.careconnect.coreapi.communications.domain;
